feat(fe): separate component and use form
